requirements.txt
asgiref==3.7.2
deprecation==2.1.0
Django==4.2.4
djangorestframework==3.14.0
img2pdf==0.4.4
lxml==4.9.3
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pdfkit==1.0.0
pikepdf==8.4.0
Pillow==10.0.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
